+++ PolynomialCheckerInterface.hpp
@@ -11,7 +11,7 @@ class PolynomialCheckerInterface
 
         // solve for needle = coeff0 + coeff1*theConst + coeff2*theConst^2 + ... (a polynomial with the given degree)
         // note that the loopRanges define indexes of coeffArray to search through
-        virtual std::vector<float*>* findHits(
+        virtual std::vector<int*>* findHits(
             const float needle,
             const float theConst,
             const int degree,
